Update instructions in README for interactive screen and container
The README file instructions on where to start the container from has been expanded. It now clarifies that it can also be started from any directory higher up the tree, and that the HTML files will be found in the /data directory or below. The note about HTML files location mapping in `docker run` command has been moved up into the main instructions.